+++ packages/core/index.js
@@ -1,0 +1,8 @@
+import dedent from 'dedent'
+
+const a = `
+  hello
+    world
+`
+
+console.log(dedent(a))