+++ packages/babel-preset-app-react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inventorjs/babel-preset-app-react
 
+## 1.2.0
+
+### Minor Changes
+
+- 9ca3719: publish new version
+
 ## 1.1.0
 
 ### Minor Changes