+++ packages/plugin-app/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @inventorjs/plugin-app
 
+## 1.6.0
+
+### Minor Changes
+
+- 9ca3719: publish new version
+
+### Patch Changes
+
+- Updated dependencies [9ca3719]
+  - @inventorjs/core@3.8.0
+
 ## 1.5.0
 
 ### Minor Changes